Simple Cache Viewer: Change all remaining boost::filesystem methods to their non-throwing overloads

+++ indra/llfilesystem/lldiskcache.cpp
@@ -71,6 +71,7 @@ void LLDiskCache::purge()
         LL_INFOS() << "Total dir size before purge is " << dirFileSize(mCacheDir) << LL_ENDL;
     }
 
+    boost::system::error_code ec;
     auto start_time = std::chrono::high_resolution_clock::now();
 
     typedef std::pair<std::time_t, std::pair<uintmax_t, std::string>> file_info_t;
@@ -81,19 +82,27 @@ void LLDiskCache::purge()
 #else
     std::string cache_path(mCacheDir);
 #endif
-    if (boost::filesystem::is_directory(cache_path))
+    if (boost::filesystem::is_directory(cache_path, ec) && !ec.failed())
     {
         // <FS:Ansariel> Optimize asset simple disk cache
         //for (auto& entry : boost::make_iterator_range(boost::filesystem::directory_iterator(cache_path), {}))
         for (auto& entry : boost::make_iterator_range(boost::filesystem::recursive_directory_iterator(cache_path), {}))
         {
-            if (boost::filesystem::is_regular_file(entry))
+            if (boost::filesystem::is_regular_file(entry, ec) && !ec.failed())
             {
                 if (entry.path().string().find(mCacheFilenamePrefix) != std::string::npos)
                 {
-                    uintmax_t file_size = boost::filesystem::file_size(entry);
+                    uintmax_t file_size = boost::filesystem::file_size(entry, ec);
+                    if (ec.failed())
+                    {
+                        continue;
+                    }
                     const std::string file_path = entry.path().string();
-                    const std::time_t file_time = boost::filesystem::last_write_time(entry);
+                    const std::time_t file_time = boost::filesystem::last_write_time(entry, ec);
+                    if (ec.failed())
+                    {
+                        continue;
+                    }
 
                     file_info.push_back(file_info_t(file_time, { file_size, file_path }));
                 }
@@ -119,7 +128,6 @@ void LLDiskCache::purge()
             action = "DELETE:";
             // <FS:Ansariel> Do not crash if we cannot delete the file for some reason
             //boost::filesystem::remove(entry.second.second);
-            boost::system::error_code ec;
             boost::filesystem::remove(entry.second.second, ec);
             if (ec.failed())
             {
@@ -321,24 +329,24 @@ void LLDiskCache::clearCache()
      * the component files but it's called infrequently so it's
      * likely just fine
      */
+    boost::system::error_code ec;
 #if LL_WINDOWS
     std::wstring cache_path(utf8str_to_utf16str(mCacheDir));
 #else
     std::string cache_path(mCacheDir);
 #endif
-    if (boost::filesystem::is_directory(cache_path))
+    if (boost::filesystem::is_directory(cache_path, ec) && !ec.failed())
     {
         // <FS:Ansariel> Optimize asset simple disk cache
         //for (auto& entry : boost::make_iterator_range(boost::filesystem::recursive_directory_iterator(cache_path), {}))
         for (auto& entry : boost::make_iterator_range(boost::filesystem::directory_iterator(cache_path), {}))
         {
-            if (boost::filesystem::is_regular_file(entry))
+            if (boost::filesystem::is_regular_file(entry, ec) && !ec.failed())
             {
                 if (entry.path().string().find(mCacheFilenamePrefix) != std::string::npos)
                 {
                     // <FS:Ansariel> Do not crash if we cannot delete the file for some reason
                     //boost::filesystem::remove(entry);
-                    boost::system::error_code ec;
                     boost::filesystem::remove(entry, ec);
                     if (ec.failed())
                     {
@@ -364,22 +372,27 @@ uintmax_t LLDiskCache::dirFileSize(const std::string dir)
      * so if performance is ever an issue, optimizing this or removing it altogether,
      * is an easy win.
      */
+    boost::system::error_code ec;
 #if LL_WINDOWS
     std::wstring dir_path(utf8str_to_utf16str(dir));
 #else
     std::string dir_path(dir);
 #endif
-    if (boost::filesystem::is_directory(dir_path))
+    if (boost::filesystem::is_directory(dir_path, ec) && !ec.failed())
     {
         // <FS:Ansariel> Optimize asset simple disk cache
         //for (auto& entry : boost::make_iterator_range(boost::filesystem::directory_iterator(dir_path), {}))
         for (auto& entry : boost::make_iterator_range(boost::filesystem::recursive_directory_iterator(dir_path), {}))
         {
-            if (boost::filesystem::is_regular_file(entry))
+            if (boost::filesystem::is_regular_file(entry, ec) && !ec.failed())
             {
                 if (entry.path().string().find(mCacheFilenamePrefix) != std::string::npos)
                 {
-                    total_file_size += boost::filesystem::file_size(entry);
+                    uintmax_t file_size = boost::filesystem::file_size(entry, ec);
+                    if (!ec.failed())
+                    {
+                        total_file_size += file_size;
+                    }
                 }
             }
         }
